move comments into drawer instead
